feat(extension): chrome.sidePanel host — tools in a docked panel (P2-3 PR-1)
