+++ extension/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to `@sfdt/extension` are documented here. Format follows [Ke
 ## [Unreleased]
 
 ### Added
+- **Docked side panel — Workspace-in-a-dock (P2-3, PR-1 of 2).** A new `chrome.sidePanel` host (`entrypoints/sidepanel`) runs the full Workspace — the same tool set, tabbed pane host, org switcher, and present-view sink — inside a docked side panel beside the Salesforce page, so your tools stay open next to the tab you're working on. Open it from the toolbar popup's new **"Open side panel"** button (a live user gesture calls `chrome.sidePanel.open({ tabId })`); the toolbar button still opens the action popup as before (no `openPanelOnActionClick` hijack). Org binding is **bind-on-open**: the panel binds to the org of the tab it was opened from (the active tab's URL), falling back to your last-used org, then an org picker — identical to the Workspace. The shared boot/layout/tool-registration is factored into a chrome-free `ui/workspace-host.ts` that both the Workspace tab and the panel import, so the two can't drift and the Workspace behaves exactly as before. Firefox parity ships too: a per-browser manifest emits `sidebar_action` (native sidebar, no extra permission) pointing at the same page. Adds exactly one new permission, **`sidePanel`** (Chrome only; ledgered, `PRIVACY.md` + store listing updated) — no new data access and no new host permission. Org-follow-across-tab-switches, auto-enable-on-Lightning, and routing the ⚡ menu into the panel are PR-2 (browser-validated). No new runtime dependency.
 - **Schema Browser — SOQL insert + export-for-prompt field selection (P2-1, PR-3 of 3).** The Schema Browser field table now integrates with two sibling tools. Each field row gains an **"Insert into query"** quick action (next to "Copy API name") that drops the field's API name into the live SOQL Runner draft — appended into the SELECT list when the runner is open, or stashed as a pending fragment and applied on the next open (mirrors the existing Saved-SOQL pending-query hand-off). The runner exposes a small imperative `insertFieldIntoDraft(field)` hook, threaded into the Schema Browser from both `content.ts` and the Workspace factory; when the hook is absent the action is hidden. For LLM prompts, the field table now **pulls back every field by API name, all pre-selected** (a labelled, keyboard-reachable checkbox per row), and you **unselect** the ones you don't want, with **Select all** / **Clear all** affordances and a live "N of M selected" count. An **"Export selected for prompt"** action emits only the chosen fields through the existing `export-for-prompt` tool — `buildSchemaMarkdown` now takes an optional field subset and falls back to the full-object export when none is supplied (the existing one-shot export path is unchanged). A per-object selection store (object API name → set of field API names) preserves each object's choices across navigation. No new runtime dependency, no new permission.
 - **Command palette — custom shortcuts (P2-2, PR-3 of 3).** You can now define your own **name → URL shortcuts** that appear in the ⚡ command palette under the **Shortcuts** category. A new **Custom shortcuts** section on the Options page manages them: add rows, edit a name/URL inline, and remove — saved through the existing single Save button. Validation rejects a **duplicate name** and a **malformed URL** (the same `z.string().url()` boundary the setting is stored behind), so a bad entry is caught on save rather than opened. Shortcuts persist as a top-level `customShortcuts` preference in `chrome.storage.local` (a global preference like theme/bridge, not a per-feature toggle). Selecting a shortcut opens its URL in a **new tab** via `window.open` (never navigating away from the Salesforce page, never `eval`), and — because it flows through the same recent-first candidate pool — a used shortcut resurfaces in the palette's **Recent** section on reopen. No new runtime dependency, no new permission.
 - **Command palette overlay + wiring (P2-2, PR-2 of 3).** The ⚡ command palette is now a real, keyboard-first overlay (`ui/command-palette.ts`) that mounts into the shared closed shadow root. A labelled search box drives a `role="listbox"` of `role="option"` rows grouped under per-category headers (Recent / This record / Features / Setup / Shortcuts / Objects); candidates and their recent-first ordering come from PR-1's `lib/palette-sources.ts` and ranking from `lib/fuzzy.ts` — nothing is re-implemented. Full keyboard contract: open focuses the search box, ArrowUp/Down move the active option via `aria-activedescendant`, Enter runs it, Esc closes and restores focus to wherever it was, Tab is trapped inside, and a click outside dismisses (shadow-safe via `composedPath()`). Opening does **zero** Salesforce API calls — features/setup/shortcuts/recents are synchronous, and the Objects category kicks off the shared describe cache only *after* first paint, filling in when it resolves (AC-1). The `open-palette` keyboard command (`Ctrl/Cmd+Shift+K`) and a new persistent **"View all features"** entry in the ⚡ menu both open it; the ⚡ menu stays the dynamic location-aware shortlist while the palette is the exhaustive searchable view (both surfaces kept). Executing a feature dispatches it, a setup/shortcut navigates, a record opens the inspector, and an object opens the schema browser when that feature is present, else the Object Manager page. The palette is itself a kill-switchable registry feature (`command-palette`, default on); when it's disabled the keyboard command falls back to the ⚡ menu. No new runtime dependency, no new permission.
