+++ extension/PRIVACY.md
@@ -74,6 +74,7 @@ The extension manifest requests the following permissions. Each is used for the 
 | `clipboardWrite` | Copy generated output to your clipboard on demand — e.g. the Flow Health Check report, SOQL results, the object schema for prompts, and the Copy-JSON actions on the Org Limits / Org Health panels |
 | `cookies` | Detect which Salesforce orgs you're logged in to by reading your existing Salesforce session cookies, so the Workspace org picker / Switch Org can target the org you choose. Cookies are read locally and never transmitted off your device. |
 | `contextMenus` | Add an optional right-click **"SFDT: Inspect this record"** item on Salesforce pages, so you can open the record inspector for the record in the page URL or a right-clicked record link. The menu only reads the URL to find a record Id; it never reads cookies or session data, and it only appears on your Salesforce hosts. Toggle it off under Options → Features. |
+| `sidePanel` | Dock the extension's own tool panel alongside Lightning, so the Workspace tools stay open beside the Salesforce page you're working on. It reads no new data and expands no host access — the panel uses the same in-browser session resolution as the Workspace tab. Chrome only; on Firefox the same panel opens through the native sidebar with no extra permission. |
 | `host_permissions: https://*.salesforce.com/*, https://*.salesforce-setup.com/*, https://*.my.salesforce.com/*, https://*.lightning.force.com/*` | Run feature scripts and call the Tooling/REST APIs on Salesforce pages of your logged-in org |
 | `host_permissions: https://*.my.salesforce.mil/*, https://*.lightning.force.mil/*` | Same support for **US Government Cloud (GovCloud)** orgs, which serve on the `.mil` top-level domain. Only the `sid` cookie of these hosts is read, locally, to resolve the org's API session — never transmitted off your device. |
 | `host_permissions: https://*.sfcrmapps.cn/*` | Same support for **Salesforce China** orgs (operated in-country on the `.sfcrmapps.cn` domain). Session resolution only; nothing leaves your device. |
