+++ extension/entrypoints/sidepanel/index.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>SFDT Panel</title>
+  </head>
+  <body>
+    <div id="sfdt-app-root"></div>
+    <script type="module" src="./main.ts"></script>
+  </body>
+</html>
